pyproject.toml
[project]
name = "code-comprehender"
version = "0.1.0"
description = "A tool for generating Java comments & generating architecture diagrams."
requires-python = ">=3.8"

dependencies = [
    "google-genai~=1.17.0",
    "graphviz~=0.20.3",
    "langchain-core~=0.3.63",
    "langchain-google-genai~=2.1.5",
    "langchain-pinecone~=0.2.6",
    "pinecone>=6.0.0,<=6.0.2",
    "python-dotenv~=1.1.0",
    "ruff~=0.11.12",
    "tree-sitter~=0.24.0",
    "tree-sitter-java~=0.23.5",
]

[tool.ruff]
line-length = 100
exclude = ["venv", ".venv"]

